find_repo: Return only the first path
Previously, if a querystring matched multiple paths, it would return all
of the possible paths, separated by string: "/path/1/ /other/path/3/".
This caused problems with the `goto` function, because the entire result
of find_repo was passed along to `cd`.

So, use `cut` to trim down to just the first path.